--- onboarding-functions/src/main/java/it/pagopa/selfcare/entity/Onboarding.java
@@ -1,113 +1,0 @@
-package it.pagopa.selfcare.entity;
-
-
-import io.quarkus.mongodb.panache.common.MongoEntity;
-import it.pagopa.selfcare.onboarding.common.OnboardingStatus;
-import org.bson.codecs.pojo.annotations.BsonId;
-import org.bson.codecs.pojo.annotations.BsonIgnore;
-import org.bson.types.ObjectId;
-
-import java.time.LocalDateTime;
-import java.time.OffsetDateTime;
-import java.util.List;
-
-
-@MongoEntity(collection="onboardings")
-public class Onboarding  {
-
-    private ObjectId id;
-
-    private String onboardingId;
-
-    private String productId;
-    private Institution institution;
-    private List<User> users;
-    private String pricingPlan;
-    private Billing billing;
-    private Boolean signContract;
-
-
-    private LocalDateTime expiringDate;
-
-    private OnboardingStatus status;
-
-    public ObjectId getId() {
-        return id;
-    }
-
-    public void setId(ObjectId id) {
-        this.id = id;
-    }
-
-    public String getProductId() {
-        return productId;
-    }
-
-    public void setProductId(String productId) {
-        this.productId = productId;
-    }
-
-    public Institution getInstitution() {
-        return institution;
-    }
-
-    public void setInstitution(Institution institution) {
-        this.institution = institution;
-    }
-
-    public List<User> getUsers() {
-        return users;
-    }
-
-    public void setUsers(List<User> users) {
-        this.users = users;
-    }
-
-    public String getPricingPlan() {
-        return pricingPlan;
-    }
-
-    public void setPricingPlan(String pricingPlan) {
-        this.pricingPlan = pricingPlan;
-    }
-
-    public Billing getBilling() {
-        return billing;
-    }
-
-    public void setBilling(Billing billing) {
-        this.billing = billing;
-    }
-
-    public Boolean getSignContract() {
-        return signContract;
-    }
-
-    public void setSignContract(Boolean signContract) {
-        this.signContract = signContract;
-    }
-
-    public LocalDateTime getExpiringDate() {
-        return expiringDate;
-    }
-
-    public void setExpiringDate(LocalDateTime expiringDate) {
-        this.expiringDate = expiringDate;
-    }
-
-    public OnboardingStatus getStatus() {
-        return status;
-    }
-
-    public void setStatus(OnboardingStatus status) {
-        this.status = status;
-    }
-
-    public String getOnboardingId() {
-        return onboardingId;
-    }
-
-    public void setOnboardingId(String onboardingId) {
-        this.onboardingId = onboardingId;
-    }
-}
