--- onboarding-functions/src/main/java/it/pagopa/selfcare/entity/Token.java
@@ -1,116 +1,0 @@
-package it.pagopa.selfcare.entity;
-
-import io.quarkus.mongodb.panache.common.MongoEntity;
-import it.pagopa.selfcare.onboarding.common.TokenType;
-import org.bson.types.ObjectId;
-
-import java.time.LocalDateTime;
-import java.time.OffsetDateTime;
-import java.util.List;
-
-
-@MongoEntity(collection="tokens")
-public class Token {
-
-    private ObjectId id;
-    private TokenType type;
-    private String productId;
-    private String checksum;
-    private String contractVersion;
-    private String contractTemplate;
-    private String contractSigned;
-    //@Indexed
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-    private LocalDateTime deletedAt;
-    private LocalDateTime activatedAt;
-
-    public ObjectId getId() {
-        return id;
-    }
-
-    public void setId(ObjectId id) {
-        this.id = id;
-    }
-
-    public TokenType getType() {
-        return type;
-    }
-
-    public void setType(TokenType type) {
-        this.type = type;
-    }
-
-    public String getProductId() {
-        return productId;
-    }
-
-    public void setProductId(String productId) {
-        this.productId = productId;
-    }
-
-    public String getChecksum() {
-        return checksum;
-    }
-
-    public void setChecksum(String checksum) {
-        this.checksum = checksum;
-    }
-
-    public String getContractVersion() {
-        return contractVersion;
-    }
-
-    public void setContractVersion(String contractVersion) {
-        this.contractVersion = contractVersion;
-    }
-
-    public String getContractTemplate() {
-        return contractTemplate;
-    }
-
-    public void setContractTemplate(String contractTemplate) {
-        this.contractTemplate = contractTemplate;
-    }
-
-    public String getContractSigned() {
-        return contractSigned;
-    }
-
-    public void setContractSigned(String contractSigned) {
-        this.contractSigned = contractSigned;
-    }
-
-    public LocalDateTime getCreatedAt() {
-        return createdAt;
-    }
-
-    public void setCreatedAt(LocalDateTime createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    public LocalDateTime getUpdatedAt() {
-        return updatedAt;
-    }
-
-    public void setUpdatedAt(LocalDateTime updatedAt) {
-        this.updatedAt = updatedAt;
-    }
-
-    public LocalDateTime getDeletedAt() {
-        return deletedAt;
-    }
-
-    public void setDeletedAt(LocalDateTime deletedAt) {
-        this.deletedAt = deletedAt;
-    }
-
-    public LocalDateTime getActivatedAt() {
-        return activatedAt;
-    }
-
-    public void setActivatedAt(LocalDateTime activatedAt) {
-        this.activatedAt = activatedAt;
-    }
-}
-
