--- onboarding-functions/src/main/java/it/pagopa/selfcare/exception/GenericOnboardingException.java
@@ -1,15 +1,0 @@
-package it.pagopa.selfcare.exception;
-
-public class GenericOnboardingException extends RuntimeException {
-
-    private final String code;
-
-    public GenericOnboardingException(String message, String code) {
-        super(message);
-        this.code = code;
-    }
-
-    public String getCode() {
-        return code;
-    }
-}
