+++ onboarding-functions/src/main/java/it/pagopa/selfcare/onboarding/entity/DataProtectionOfficer.java
@@ -1,0 +1,32 @@
+package it.pagopa.selfcare.onboarding.entity;
+
+
+public class DataProtectionOfficer {
+    private String address;
+    private String email;
+    private String pec;
+
+    public String getAddress() {
+        return address;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPec() {
+        return pec;
+    }
+
+    public void setPec(String pec) {
+        this.pec = pec;
+    }
+}
