+++ onboarding-functions/src/main/java/it/pagopa/selfcare/onboarding/repository/TokenRepository.java
@@ -1,0 +1,9 @@
+package it.pagopa.selfcare.onboarding.repository;
+
+import io.quarkus.mongodb.panache.PanacheMongoRepository;
+import it.pagopa.selfcare.onboarding.entity.Token;
+import jakarta.enterprise.context.ApplicationScoped;
+
+@ApplicationScoped
+public class TokenRepository implements PanacheMongoRepository<Token> {
+}
