+++ onboarding-functions/src/main/java/it/pagopa/selfcare/onboarding/service/NotificationService.java
@@ -1,0 +1,7 @@
+package it.pagopa.selfcare.onboarding.service;
+
+
+public interface NotificationService {
+
+    void sendMailWithContract(String onboardingId, String filenameContract, String destination, String name, String username, String productName, String token);
+}
