--- onboarding-functions/src/main/java/it/pagopa/selfcare/repository/TokenRepository.java
@@ -1,9 +1,0 @@
-package it.pagopa.selfcare.repository;
-
-import io.quarkus.mongodb.panache.PanacheMongoRepository;
-import it.pagopa.selfcare.entity.Token;
-import jakarta.enterprise.context.ApplicationScoped;
-
-@ApplicationScoped
-public class TokenRepository implements PanacheMongoRepository<Token> {
-}
