--- onboarding-functions/src/main/java/it/pagopa/selfcare/service/NotificationService.java
@@ -1,7 +1,0 @@
-package it.pagopa.selfcare.service;
-
-
-public interface NotificationService {
-
-    void sendMailWithContract(String onboardingId, String filenameContract, String destination, String name, String username, String productName, String token);
-}
