+++ onboarding-functions/src/test/java/it/pagopa/selfcare/OnboardingFunctionsTest.java
@@ -7,6 +7,7 @@ import com.microsoft.azure.functions.HttpStatus;
 import com.microsoft.durabletask.DurableTaskClient;
 import com.microsoft.durabletask.azurefunctions.DurableClientContext;
 import io.quarkus.test.junit.QuarkusTest;
+import it.pagopa.selfcare.onboarding.OnboardingFunctions;
 import org.mockito.invocation.InvocationOnMock;
 import org.mockito.stubbing.Answer;
